Replace SerializableTimeOnly with TimeOnlyAttributes
SerializableTimeOnly and SerializableTimeSpan were wrappers for serializing a TimeOnly and TimeSpan field respectively. However, these types can be serialized with a simple long, so a different approach has been taken. The new way of serializing a TimeOnly variable is like this:

[TimeOnly] public long timeOnly;

This draws a TimeOnly drawer in the inspector while using a long field under the hood. If you want, you can use ISerializationCallbackReceiver to serialize a real TimeOnly value like so:

[SerializeField, TimeOnly] private long _timeOnly;
public TimeOnly timeOnly;

void OnBeforeSerialize()
{
    _timeOnly = timeOnly.Ticks;
}

void OnAfterDeserialize()
{
    timeOnly = new(_timeOnly);
}

The same works for TimeSpan.