Add UIElements implementation
TimeOnlyField and TimeSpanField are proper implementations for TimeOnly and TimeSpan in the inspector. The field can be dragged in order to increase or decrease the time field. Naturally, inputting a value via text still works as well.

Additionally, the TimeOnlyAttribute and TimeSpanAttribute have been simplified. They can no longer take any additional values and instead always display the foldout. In most cases, the UIElement implementation will be used, so this is not a real issue.