Change version to 0.9.0-pre (#1)
* Add UIElements implementation

TimeOnlyField and TimeSpanField are proper implementations for TimeOnly and TimeSpan in the inspector. The field can be dragged in order to increase or decrease the time field. Naturally, inputting a value via text still works as well.

Additionally, the TimeOnlyAttribute and TimeSpanAttribute have been simplified. They can no longer take any additional values and instead always display the foldout. In most cases, the UIElement implementation will be used, so this is not a real issue.

* Make ClockVolume Optional

ClockVolume put a dependency on the render pipeline core library, and for games to use a render pipeline in order to work with time. This update makes the clock volume completely optional: handy if you want it, but not required.

* Remove Temporal scripts

The temporal method was experimental and a new method relying on cycles is under way.

* Rename Shader Includes and Methods

_UnityClock_Interpolant has been renamed to _UnityClock_Day.
_UnityClock_PingPong has been renamed to _UnityClock_Midday.

Clock.PingPong has been renamed to Clock.Midday.
Clock.Day was added for consistency.

* Add Cycles

Clock now has a playable graph to connect to. This can be done via cycles. Cycles are Playables that cycle back to their starting point based on the time of day. This provides a grounded implementation for adding and removing functionality to the clock in a lightweight manner.

* Completely rework the Clock Tools

The clock tools are now simpler, and use UIElements. They hook into the Clock graph to sample the current scene in a very easy manner, allowing for anything that gets added to the Clock graph to be displayed by the Clock tools.

* Add ClockInstance

Provide a component that can easily be added to a scene to set time in a simple, lightweight manner.

* Remove ClockVolume

It was a failed experiment: a different solution is required in order to blend time based on your position.

* Small fixes

- CycleAnimator needs to call evaluate after it sets the time of its playables.
- Clock doesn't need to call Play and Stop since the time is updated manually.

* Remove OnGUI drawers

These are old and unnecessary since the CreatePropertyGUI implementation.

* Change version to 0.9.0-pre

Changed the version in package.json.
Updated the CHANGELOG.md.
Updated the README.md.